UICIRC-121 - added new validation approach. Changed files structure
